Restructure Redux store directory tree
Following the plan outlined in #687, restructure the Redux store as
follows:

 - Move the entry point from `services/annotation-ui.js` to `store/index.js`
 - Move the modules which define bundles of reducers, action creators
   and selectors to `store/modules/{module-name}.js`
 - Move related utility code to `store/{module-name}.js`
