+++ src/sidebar/store/debug-middleware.js
