+++ src/sidebar/store/test/index-test.js
@@ -2,7 +2,7 @@
 
 var immutable = require('seamless-immutable');
 
-var annotationUIFactory = require('../annotation-ui');
+var annotationUIFactory = require('../index');
 var annotationFixtures = require('../../test/annotation-fixtures');
 var metadata = require('../../annotation-metadata');
 var unroll = require('../../../shared/test/util').unroll;
