+++ src/sidebar/store/util.js
